Add CustomNonceModal component
